+++ src/app/layout.tsx
@@ -1,3 +1,21 @@
+import "@fontsource-variable/montserrat";
+import "@fontsource-variable/source-code-pro";
+import "@fontsource-variable/roboto-slab";
+import "@fontsource/lily-script-one";
+import "styles/globals.css";
+import posthog from "posthog-js";
+
+if (typeof window !== "undefined") {
+    // checks that we are client-side
+    posthog.init(process.env.NEXT_PUBLIC_POSTHOG_KEY as string, {
+        api_host: "/ingest",
+        ui_host: "https://eu.posthog.com",
+        loaded: (posthog) => {
+            if (process.env.NODE_ENV === "development") posthog.debug(); // debug mode in development
+        },
+    });
+}
+
 export const metadata = {
     title: "MyBPHS",
     description: "MyBPHS",
