+++ src/components/BigLinkButton.tsx
@@ -1,6 +1,8 @@
+"use client";
+
 import React, { useState } from "react";
 import { motion } from "framer-motion";
-import { useRouter } from "next/router";
+import { useRouter } from "next/navigation";
 
 function BigLinkButton({
     title,
